windows_managers_view/requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asyncio==4.0.0
attrs==25.4.0
cabarchive==0.2.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cx_Freeze==8.4.1
cx_Logging==3.2.1
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.121.1
ffmpeg-python==0.2.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hikvisionapi==0.3.2
httpcore==1.0.9
httpx==0.28.1
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
lief==0.17.0
MarkupSafe==3.0.3
matplotlib==3.10.7
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pefile==2023.2.7
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
polars==1.35.2
polars-runtime-32==1.35.2
psutil==7.1.3
py-cpuinfo==9.0.0
py2exe==0.14.0.0
PyAutoGUI==0.9.54
pybind11==3.0.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyGetWindow==0.0.9
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
scipy==1.15.3
shiboken6==6.10.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
striprtf==0.0.29
sympy==1.14.0
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
uuid==1.30
uvicorn==0.38.0
websockets==15.0.1
xml-python==0.4.3
xmltodict==1.0.2
